Widen Scala set element types for mixed-width ints
Bugbot caught that Scala's typed-opener config was missing the
``wide_int_type`` mapping added in the earlier widening commit.
Without it ``WideInt`` fell back to ``int_type="Int"``, producing
``Set[Int](1, 1099511627776L)`` — a Long literal pinned to an Int
collection type, which Scala rejects at compile time.  Map
``WideInt`` to ``Long`` so the annotation becomes ``Set[Long]`` and
Scala's compile-time literal widening promotes the Int literals to
Long.

+++ tests/integration/cases/set_mixed_int_widths/Scala.scala
@@ -1,5 +1,5 @@
 object Fixture_set_mixed_int_widths_Scala {
-val my_data = Set[Int](
+val my_data = Set[Long](
     1,
     1099511627776L,
 )
