+++ tests/integration/cases/set_mixed_int_widths/Scala_declaration_style_var.scala
@@ -1,5 +1,5 @@
 object Fixture_set_mixed_int_widths_Scala_declaration_style_var {
-var my_data = Set[Int](
+var my_data = Set[Long](
     1,
     1099511627776L,
 )
